+++ pkg/asset/directoriesreader.go
@@ -1,4 +1,4 @@
-// Copyright Contributors to the Open Cluster Management project
+// Copyright Red Hat
 
 package asset
 
@@ -7,9 +7,7 @@ import (
 	"io/ioutil"
 	"os"
 	"path/filepath"
-	"strings"
 
-	"github.com/ghodss/yaml"
 	"k8s.io/klog/v2"
 )
 
@@ -86,50 +84,4 @@ func (r *YamlFileReader) AssetNames(excluded []string) ([]string, error) {
 		}
 	}
 	return files, nil
-}
-
-//ToJSON converts to JSON
-func (*YamlFileReader) ToJSON(
-	b []byte,
-) ([]byte, error) {
-	b, err := yaml.YAMLToJSON(b)
-	if err != nil {
-		klog.Errorf("err:%s\nyaml:\n%s", err, string(b))
-		return nil, err
-	}
-	return b, nil
-}
-
-func (r *YamlFileReader) ExtractAssets(prefix, dir string, excluded []string) error {
-	assetNames, err := r.AssetNames(excluded)
-	if err != nil {
-		return err
-	}
-	for _, assetName := range assetNames {
-		if !strings.HasPrefix(assetName, prefix) {
-			continue
-		}
-		relPath, err := filepath.Rel(prefix, assetName)
-		if err != nil {
-			return err
-		}
-		path := filepath.Join(dir, relPath)
-
-		if relPath == "." {
-			path = filepath.Join(dir, filepath.Base(assetName))
-		}
-		err = os.MkdirAll(filepath.Dir(path), os.FileMode(0700))
-		if err != nil {
-			return err
-		}
-		data, err := r.Asset(assetName)
-		if err != nil {
-			return err
-		}
-		err = ioutil.WriteFile(path, data, os.FileMode(0600))
-		if err != nil {
-			return err
-		}
-	}
-	return nil
 }